Add get all votes page and edit file names
