+++ frontend/src/app/vote/get_all/page.tsx
@@ -1,0 +1,11 @@
+import Votes from '@/components/VoteList'
+
+function page() {
+  return (
+    <div className="flex flex-col items-center justify-center min-h-screen bg-gray-100">
+      <Votes />
+    </div>
+  )
+}
+
+export default page
